Move possession tracker to sensor fusion
