+++ src/software/world/team_types.h
@@ -11,3 +11,8 @@ MAKE_ENUM(TeamSide, FRIENDLY, ENEMY);
  * The possible team colours of SSL robots
  */
 MAKE_ENUM(TeamColour, YELLOW, BLUE)
+
+/**
+ * Indicates which team has possession of the ball
+ */
+MAKE_ENUM(TeamPossession, FRIENDLY_TEAM, ENEMY_TEAM)
