+++ src/software/world/world.cpp
@@ -14,7 +14,7 @@ World::World(const Field &field, const Ball &ball, const Team &friendly_team,
       // Store a small buffer of previous referee commands so we can filter out noise
       referee_command_history_(REFEREE_COMMAND_BUFFER_SIZE),
       referee_stage_history_(REFEREE_COMMAND_BUFFER_SIZE),
-      team_with_possesion_(TeamSide::ENEMY)
+      team_with_possession_(TeamPossession::FRIENDLY_TEAM)
 {
     updateTimestamp(getMostRecentTimestampFromMembers());
 }
@@ -164,12 +164,12 @@ const RefereeStage &World::getRefereeStage() const
     return current_referee_stage_;
 }
 
-void World::setTeamWithPossession(TeamSide team_with_possesion)
+void World::setTeamWithPossession(TeamPossession team_with_possesion)
 {
-    team_with_possesion_ = team_with_possesion;
+    team_with_possession_ = team_with_possesion;
 }
 
-TeamSide World::getTeamWithPossession() const
+TeamPossession World::getTeamWithPossession() const
 {
-    return team_with_possesion_;
+    return team_with_possession_;
 }
